Refactor Aij -> AIJExt
